+++ Modules/Project/Tarun Moradiya/React/rx-training/src/components/Tech.js
@@ -1,11 +1,14 @@
-import React, { useRef, useState, useContext } from "react";
+import React, { useRef, useState, useContext, useEffect } from "react";
 import { Link, useHistory } from "react-router-dom";
 import loadingGif from "../images/gif/loading-arrow.gif";
 import { TechContext } from "../contexts/techContext";
 import AddModule from "./AddModule";
 import { AuthContext } from "../contexts/authContext";
-import { Button } from "react-bootstrap";
+import { Button, Accordion, Card } from "react-bootstrap";
+import { PlanContext } from "../contexts/planContext";
 import { ModuleContext } from "../contexts/moduleContext";
+import SubItem from "./SubItem";
+import styled from "styled-components";
 
 export default function Tech(props) {
   const [editItem, setEditItem] = useState(false);
@@ -17,10 +20,14 @@ export default function Tech(props) {
   const [show, setShow] = useState(false);
   const [showForm, setShowForm] = useState(false);
 
-  const { modules } = useContext(ModuleContext);
   const history = useHistory();
   const { techName, setTechName, updateTech, deleteTech } =
     useContext(TechContext);
+
+  const [techModules, setTechModules] = useState([]);
+  const [techPlans, setTechPlans] = useState([]);
+  const { modules } = useContext(ModuleContext);
+  const { plans } = useContext(PlanContext);
 
   //show and hide modal
   const handleClose = () => {
@@ -34,9 +41,14 @@ export default function Tech(props) {
 
   //delete tech
   const handleDeleteBtn = async () => {
-    await setLoading(true);
-    await deleteTech(tech._id);
-    await setLoading(false);
+    const del = window.confirm(
+      `Are You Sure, You Want To Delete ${tech.name} ?`
+    );
+    if (del) {
+      await setLoading(true);
+      await deleteTech(tech._id);
+      await setLoading(false);
+    }
   };
 
   //show edit tech form
@@ -66,11 +78,24 @@ export default function Tech(props) {
     }
   };
 
-  const getModule = async () => {
-    const firstModule = modules.find((module) => module.tech === tech._id);
-    if (!firstModule) history.push("error");
-    else history.push("/modules/" + firstModule._id);
+  // const getModule = async () => {
+  //   const firstModule = modules.find((module) => module.tech === tech._id);
+  //   if (!firstModule) history.push("error");
+  //   else history.push("/modules/" + firstModule._id);
+  // };
+
+  const getData = async () => {
+    const filteredModules = modules.filter(
+      (module) => module.tech === tech._id
+    );
+    await setTechModules(filteredModules);
+    const filteredPlans = plans.filter((plan) => plan.tech === tech._id);
+    await setTechPlans(filteredPlans);
   };
+
+  useEffect(() => {
+    getData();
+  }, [plans, modules]);
 
   if (loading) {
     return (
@@ -85,7 +110,7 @@ export default function Tech(props) {
   }
 
   return (
-    <li className="list-group-item bg-light border-info">
+    <CustomCard className="list-group-item bg-light border-info clearfix">
       {editItem ? (
         <form onSubmit={handleEdit} className="form-inline my-2 ">
           <div className="input-group w-100">
@@ -106,45 +131,95 @@ export default function Tech(props) {
         </form>
       ) : (
         <>
-          <a
+          <Card.Header className="header bg-light pb-3 clearfix">
+            <Accordion.Toggle as={A} variant="link" eventKey={tech._id}>
+              {tech.name}
+            </Accordion.Toggle>
+
+            {/* <a
             style={{ cursor: "pointer" }}
             onClick={getModule}
             className="list-group-item-action text-info"
           >
             {tech.name}
-          </a>
-          {isAdminUser === true && (
-            <>
-              <button
-                onClick={handleShow}
-                className="btn btn-warning float-right ml-2"
-              >
-                <i className="fas fa-plus"></i>
-              </button>
-              {showForm && (
-                <AddModule
-                  id="add-tech-type"
-                  show={show}
-                  handleClose={handleClose}
-                  techId={tech._id}
+          </a> */}
+            {isAdminUser === true && (
+              <div className="float-right">
+                <button
+                  onClick={handleShow}
+                  className="btn btn-warning float-right ml-2"
+                >
+                  <i className="fas fa-plus"></i>
+                </button>
+                {showForm && (
+                  <AddModule
+                    id="add-tech-type"
+                    show={show}
+                    handleClose={handleClose}
+                    techId={tech._id}
+                  />
+                )}
+                <button
+                  onClick={handleDeleteBtn}
+                  className="btn btn-danger float-right ml-2"
+                >
+                  <i className="fas fa-trash"></i>
+                </button>
+                <button
+                  onClick={handleEditBtn}
+                  className="btn btn-primary float-right ml-2"
+                >
+                  <i className="fas fa-pen"></i>
+                </button>
+              </div>
+            )}
+          </Card.Header>
+          <Accordion.Collapse eventKey={tech._id}>
+            <Card.Body className="body">
+              {techModules.map((module) => (
+                <SubItem
+                  key={module._id}
+                  item={module}
+                  path={`/modules/${module._id}`}
                 />
-              )}
-              <button
-                onClick={handleDeleteBtn}
-                className="btn btn-danger float-right ml-2"
-              >
-                <i className="fas fa-trash"></i>
-              </button>
-              <button
-                onClick={handleEditBtn}
-                className="btn btn-primary float-right ml-2"
-              >
-                <i className="fas fa-pen"></i>
-              </button>
-            </>
-          )}
+              ))}
+              {techPlans.map((plan) => (
+                <SubItem
+                  key={plan._id}
+                  item={plan}
+                  path={`/plans/${plan._id}`}
+                />
+              ))}
+            </Card.Body>
+          </Accordion.Collapse>
         </>
       )}
-    </li>
+    </CustomCard>
   );
 }
+
+const A = styled.a`
+  :hover {
+    cursor: pointer;
+  }
+`;
+
+const CustomCard = styled.div`
+  padding-bottom: 0;
+  a {
+    text-decoration: none;
+    color: #17a2b8;
+  }
+  .header {
+    border: 0;
+    padding: 0.1rem;
+  }
+  .body {
+    border: 0;
+    border-top: #17a2b8 1px solid;
+    padding: 0.2rem 0;
+  }
+  li {
+    padding: 0.5rem;
+  }
+`;
